pyproject.toml
[project]
name = "krddevquizbot"
version = "0.1.0"
description = ""
authors = [
    {name = "brunql", email = "brunql@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.10"

dependencies = [
    "python-telegram-bot[job-queue]>=20.7",
]

[project.urls]
Homepage = "https://t.me/krddevdays_chat"
Repository = "https://github.com/krddevdays/quizbot"

[project.scripts]
pdm = "pdm.core:main"

[tool.pdm.scripts]
_.env_file = ".env"
bot = "python -m krddevquizbot"  

# dev run: nodemon -e py --exec pdm run bot
